Merge analyze into backend

See merge request s08-bigdata-dist-sub2/S08P22E205!237

+++ backend/analyze/src/main/java/com/trendflow/analyze/analyze/dto/vo/CompareInfoVo.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CompareInfoVo {
     private String type;
-    private Integer changed;
+    private Double changed;
 }
