+++ backend/analyze/src/main/java/com/trendflow/analyze/analyze/entity/SentimentCount.java
@@ -3,7 +3,7 @@ package com.trendflow.analyze.analyze.entity;
 import java.time.LocalDate;
 
 public interface SentimentCount {
-    Double getScore();
+    Long getScore();
     Long getCount();
     LocalDate getRegDt();
 }
